Extract account settings into standalone route
